feat : revised editor's location, add editor's constants
